+++ resources/views/layouts/menu.blade.php
@@ -6,19 +6,19 @@
         </a>
     </li>
     <li class="nav-item">
-        <a class="nav-link" href="{{ route('users.index') }}">
+        <a class="nav-link" href="{{ route('laravel-iam.users.index') }}">
         <span data-feather="users"></span>
         Users
         </a>
     </li>
     <li class="nav-item">
-        <a class="nav-link" href="{{ route('roles.index') }}">
+        <a class="nav-link" href="{{ route('laravel-iam.roles.index') }}">
         <span data-feather="bar-chart-2"></span>
         Roles
         </a>
     </li>
     <li class="nav-item">
-        <a class="nav-link" href="{{ route('permissions.index') }}">
+        <a class="nav-link" href="{{ route('laravel-iam.permissions.index') }}">
         <span data-feather="layers"></span>
         Permissions
         </a>