+++ resources/views/permissions/index.blade.php
@@ -20,11 +20,11 @@
                 <tr>
                     <td>{{ $permission->name }}</td> 
                     <td>                   
-                        <a href="{{ route('permissions.edit', $permission->id) }}" 
+                        <a href="{{ route('laravel-iam.permissions.edit', $permission->id) }}" 
                             class="pull-left" style="margin-right: 3px;">
                             <span data-feather="edit"></span>
                         </a>
-                        <a href="{{ route('permissions.destroy', $permission->id) }}" 
+                        <a href="{{ route('laravel-iam.permissions.destroy', $permission->id) }}" 
                                 data-method="delete" class="pull-left" style="margin-right: 3px;">
                             <span data-feather="trash"></span>
                         </a> 
@@ -35,7 +35,7 @@
         </table>
     </div>
 
-    <a href="{{ route('permissions.create') }}" class="btn btn-success">Add Permission</a>
+    <a href="{{ route('laravel-iam.permissions.create') }}" class="btn btn-success">Add Permission</a>
 
 </div>