+++ resources/views/users/edit.blade.php
@@ -10,7 +10,7 @@
 
     @include ('laraveliam::errors.list')
 
-    {{ Form::model($user, array('route' => array('users.update', $user->id), 'method' => 'PUT')) }} {{-- Form model binding to automatically populate our fields with user data --}}
+    {{ Form::model($user, array('route' => array('laravel-iam.users.update', $user->id), 'method' => 'PUT')) }} {{-- Form model binding to automatically populate our fields with user data --}}
 
     <div class="form-group">
         {{ Form::label('name', 'Name') }}