+++ src/Http/routes.php
@@ -1,7 +1,13 @@
 <?php
 Route::get('/', 'HomeController@index')->name('laravel-iam.dashboard');
 Route::get('dashboard', 'HomeController@index')->name('laravel-iam.dashboard');
-Route::resource('users', 'UserController');
-Route::resource('roles', 'RoleController');
-Route::resource('permissions', 'PermissionController');
+Route::resource('users', 'UserController', [
+    'as' => 'laravel-iam'
+]);
+Route::resource('roles', 'RoleController', [
+    'as' => 'laravel-iam'
+]);
+Route::resource('permissions', 'PermissionController', [
+    'as' => 'laravel-iam'
+]);
 Route::get('impersonate/{user}', 'ImpersonationController')->name('impersonate');